requirements.txt
thefuzz
python-Levenshtein
pytest
telethon
pytest_asyncio
python-dotenv
openai

